+++ src/video/GLUtil.cc
@@ -260,10 +260,10 @@ void LineTexture::draw(
 	int screenR = screenL + width;
 	bind();
 	glBegin(GL_QUADS);
-	glTexCoord2f(texL, 1.0f); glVertex2i(screenL, screenY);
-	glTexCoord2f(texR, 1.0f); glVertex2i(screenR, screenY);
-	glTexCoord2f(texR, 0.0f); glVertex2i(screenR, screenY + height);
-	glTexCoord2f(texL, 0.0f); glVertex2i(screenL, screenY + height);
+	glTexCoord2f((GLfloat)texL, 1.0f); glVertex2i(screenL, screenY);
+	glTexCoord2f((GLfloat)texR, 1.0f); glVertex2i(screenR, screenY);
+	glTexCoord2f((GLfloat)texR, 0.0f); glVertex2i(screenR, screenY + height);
+	glTexCoord2f((GLfloat)texL, 0.0f); glVertex2i(screenL, screenY + height);
 	glEnd();
 }
 
@@ -336,7 +336,7 @@ void StoredFrame::drawBlend(
 	glEnable(GL_BLEND);
 	glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 	// RGB come from texture, alpha comes from fragment colour.
-	glColor4f(1.0, 0.0, 0.0, alpha);
+	glColor4f(1.0, 0.0, 0.0, (GLfloat)alpha);
 	draw(offsetX, offsetY, width, height);
 	glDisable(GL_BLEND);
 }
